refactor: renaming and improving code
